backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5
Celery==5.3.1
redis==4.6.0
Pillow==10.0.0
python-dotenv==1.0.0
requests==2.31.0
PyJWT==2.8.0
Werkzeug==2.3.7
python-socketio==5.8.0
eventlet==0.33.3
cryptography==41.0.4